requirements.txt
lmstudio
requests
Pillow
numpy
scipy
torch
opencv-python